pyproject.toml
[project]
name = "rag-fixed-income-covenants"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.42.49",
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-text-splitters>=1.1.0",
    "pandas>=3.0.0",
    "pypdf>=6.7.0",
    "python-dotenv>=1.2.1",
    "python-magic-bin>=0.4.14",
    "sentence-transformers>=5.2.2",
    "unstructured>=0.20.2",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "pytest>=9.0.2",
    "ruff>=0.15.1",
]
